+++ src/common-suggest/index.js
@@ -1,2 +1,3 @@
 // @flow
 export { CommonSuggest } from "./common-suggest";
+export { SuggestionsOmniField } from "./suggestions-omni-field";